New modular layout successfully connected.
