+++ views/index.handlebars
@@ -52,9 +52,9 @@
 
 </body>
 
-<script type="text/javascript" src="/assets/js/playerInfo.js"></script>
-<script type="text/javascript" src="/assets/js/cms.js"></script>
-<script>
+<script type="module" src="./assets/js/playerInfo.js"></script>
+<script type="module" src="./assets/js/cms.js"></script>
+{{!-- <script>
     $("#submit").on("click", function () {
         playerInformation();
         $("#SearchPlayer").val('');
@@ -64,7 +64,7 @@
     $("#clear").on("click", function () {
         cardClear();
     });
-</script>
+</script> --}}
 
 {{!-- <script>
     $("#VDeck").on("click", function () {
