+++ views/layouts/main.handlebars
@@ -15,5 +15,6 @@
   </body>
 
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-  <script type="text/javascript" src="/assets/js/cms.js"></script>
+  {{!-- <script type="text/javascript" src="../assets/js/cms.js"></script>
+  <script type="module" src="../assets/js/playerinfo.js"></script> --}}
 </html>
